add routes admin

+++ src/app/api/account/[id]/route.ts
@@ -1,7 +1,8 @@
 import prisma from "@/lib/prisma";
 import { validateUpdateAccount } from "@/lib/validations/account";
-import { validateUpdateBank } from "@/lib/validations/bank";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
+import { authOptions } from "../../auth-options";
 
 export async function GET(
   _: NextRequest,
@@ -28,6 +29,10 @@ export async function PATCH(
   { params }: { params: { id: string } }
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    if (session.user.role !== "ADMIN") {
+      return NextResponse.json({ error: "forbidden" }, { status: 500 });
+    }
     const account_id = Number(params.id);
     const account_info = await req.json();
     const validatedAccount = validateUpdateAccount(account_info);
@@ -52,6 +57,10 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    if (session.user.role !== "ADMIN") {
+      return NextResponse.json({ error: "forbidden" }, { status: 500 });
+    }
     const account_id = Number(params.id);
     await prisma.account.delete({
       where: { id: account_id },
