+++ src/app/api/product/[id]/route.ts
@@ -1,6 +1,8 @@
 import { ProductModel } from "@/models/mysql/product-model";
 import { ProductController } from "@/service/products";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
+import { authOptions } from "../../auth-options";
 
 const productController = new ProductController({ productModel: ProductModel });
 
@@ -24,6 +26,10 @@ export async function PATCH(
   { params }: { params: { id: string } }
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    if (session.user.role !== "ADMIN") {
+      return NextResponse.json({ error: "forbidden" }, { status: 500 });
+    }
     const product = await productController.update({ req, params });
     return NextResponse.json(product);
   } catch (error: any) {
@@ -39,6 +45,10 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    if (session.user.role !== "ADMIN") {
+      return NextResponse.json({ error: "forbidden" }, { status: 500 });
+    }
     await productController.delete({ params });
     return NextResponse.json({ message: "Delete product" });
   } catch (error: any) {
